pyproject.toml
[project]
name = "convlab"
version = "3.0.2b"
description = "An Open-source Dialog System Toolkit"
authors = [{ name = "convlab" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.6.0,<2.9.0",
    "langgraph>=0.5.4",
    "langchain-community>=0.3.9",
    "langchain-openai>=0.3.7",
    "openpipe-art[langgraph,plotting,skypilot]==0.4.11",
    "boto3>=1.39.4",
    "embeddings>=0.0.8",
    "fuzzywuzzy>=0.18.0",
    "gitpython>=3.1.44",
    "litellm>=1.74.0.post1",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "torchao<=0.12",
    "python-dotenv>=1.0.1",
    "azure-identity>=1.21.0",
    "sentence-transformers>=3.0.0,<5.0.0",
    "rich>=13.0.0",
    "langgraph-cli[inmem]>=0.3.1",
    "langsmith>=0.3.37",
    "ipykernel>=6.29.5",
    "supabase>=2.15.3",
    "openai>=1.61.0",
    "pandas>=2.3.1",
    "polars>=1.33.1",
    "sacrebleu>=2.5.1",
    "tensorboardx>=2.6.4",
    "langchain>=0.3.27",
]

[dependency-groups]
dev = ["pytest>=8.4.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
